some further refactor based on feedback.
